requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
sqlalchemy==2.0.30
python-dotenv==1.0.1
vaderSentiment==3.3.2
spacy==3.7.4
scikit-learn==1.5.0
pytest==8.2.0
httpx==0.27.0
pydantic==2.7.1
langdetect==1.0.9
cachetools==5.3.3
python-dateutil==2.9.0
